+++ tomcat/src/main/java/org/apache/HomeRequestHandler.java
@@ -2,6 +2,10 @@ package org.apache;
 
 import java.io.IOException;
 
+import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+
 public class HomeRequestHandler implements RequestHandler {
 
 	private static final String URI_PATTERN = "/";
