+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequestBody.java
@@ -1,4 +1,4 @@
-package org.apache;
+package org.apache.coyote.http11;
 
 import java.util.Map;
