+++ tomcat/src/test/java/org/apache/coyote/http11/HttpHeaderTest.java
@@ -1,7 +1,8 @@
-package org.apache;
+package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.*;
 
+import org.apache.coyote.http11.HttpHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
