+++ tomcat/src/test/java/org/apache/coyote/http11/HttpMethodTest.java
@@ -1,8 +1,8 @@
-package org.apache;
+package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
+import org.apache.coyote.http11.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
