+++ tomcat/src/test/java/org/apache/coyote/http11/HttpRequestBodyTest.java
@@ -1,7 +1,8 @@
-package org.apache;
+package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.*;
 
+import org.apache.coyote.http11.HttpRequestBody;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
